+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Utils.java
@@ -4,45 +4,26 @@ package com.starrocks.sql.optimizer;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
-import com.google.common.collect.Sets;
-import com.starrocks.analysis.JoinOperator;
 import com.starrocks.catalog.Column;
-import com.starrocks.catalog.Database;
 import com.starrocks.catalog.IcebergTable;
 import com.starrocks.catalog.KeysType;
 import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.MaterializedIndex;
-import com.starrocks.catalog.MaterializedView;
-import com.starrocks.catalog.MvId;
 import com.starrocks.catalog.OlapTable;
 import com.starrocks.catalog.Partition;
 import com.starrocks.catalog.ScalarType;
 import com.starrocks.catalog.Table;
 import com.starrocks.catalog.Type;
-import com.starrocks.common.Pair;
 import com.starrocks.connector.iceberg.cost.IcebergTableStatisticCalculator;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
-import com.starrocks.sql.analyzer.Analyzer;
-import com.starrocks.sql.ast.QueryRelation;
-import com.starrocks.sql.ast.QueryStatement;
-import com.starrocks.sql.ast.StatementBase;
-import com.starrocks.sql.optimizer.base.ColumnRefFactory;
-import com.starrocks.sql.optimizer.base.ColumnRefSet;
-import com.starrocks.sql.optimizer.base.PhysicalPropertySet;
-import com.starrocks.sql.optimizer.operator.AggType;
 import com.starrocks.sql.optimizer.operator.Operator;
 import com.starrocks.sql.optimizer.operator.OperatorType;
-import com.starrocks.sql.optimizer.operator.logical.LogicalAggregationOperator;
-import com.starrocks.sql.optimizer.operator.logical.LogicalFilterOperator;
 import com.starrocks.sql.optimizer.operator.logical.LogicalHiveScanOperator;
 import com.starrocks.sql.optimizer.operator.logical.LogicalHudiScanOperator;
 import com.starrocks.sql.optimizer.operator.logical.LogicalIcebergScanOperator;
 import com.starrocks.sql.optimizer.operator.logical.LogicalJoinOperator;
 import com.starrocks.sql.optimizer.operator.logical.LogicalOlapScanOperator;
-import com.starrocks.sql.optimizer.operator.logical.LogicalOperator;
-import com.starrocks.sql.optimizer.operator.logical.LogicalProjectOperator;
 import com.starrocks.sql.optimizer.operator.logical.LogicalScanOperator;
 import com.starrocks.sql.optimizer.operator.scalar.BinaryPredicateOperator;
 import com.starrocks.sql.optimizer.operator.scalar.CastOperator;
@@ -50,13 +31,7 @@ import com.starrocks.sql.optimizer.operator.scalar.ColumnRefOperator;
 import com.starrocks.sql.optimizer.operator.scalar.CompoundPredicateOperator;
 import com.starrocks.sql.optimizer.operator.scalar.ConstantOperator;
 import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
-import com.starrocks.sql.optimizer.operator.scalar.ScalarOperatorVisitor;
-import com.starrocks.sql.optimizer.rewrite.ScalarOperatorRewriter;
-import com.starrocks.sql.optimizer.rule.RuleSetType;
 import com.starrocks.sql.optimizer.statistics.ColumnStatistic;
-import com.starrocks.sql.optimizer.transformer.LogicalPlan;
-import com.starrocks.sql.optimizer.transformer.RelationTransformer;
-import com.starrocks.sql.parser.ParsingException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -69,10 +44,8 @@ import java.util.BitSet;
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
@@ -562,326 +535,5 @@ public class Utils {
         }
 
         root.getChildren().forEach(child -> collect(child, clazz, output));
-    }
-
-    public static Set<MaterializedView> getRelatedMvs(int maxLevel, List<Table> tablesToCheck) {
-        Set<MaterializedView> mvs = Sets.newHashSet();
-        getRelatedMvs(maxLevel, 0, tablesToCheck, mvs);
-        return mvs;
-    }
-
-    public static void getRelatedMvs(int maxLevel, int currentLevel, List<Table> tablesToCheck, Set<MaterializedView> mvs) {
-        if (currentLevel >= maxLevel) {
-            return;
-        }
-        Set<MvId> newMvIds = Sets.newHashSet();
-        for (Table table : tablesToCheck) {
-            Set<MvId> mvIds = table.getRelatedMaterializedViews();
-            if (mvIds != null && !mvIds.isEmpty()) {
-                newMvIds.addAll(mvIds);
-            }
-        }
-        if (newMvIds.isEmpty()) {
-            return;
-        }
-        List<Table> newMvs = Lists.newArrayList();
-        for (MvId mvId : newMvIds) {
-            Database db = GlobalStateMgr.getCurrentState().getDb(mvId.getDbId());
-            if (db == null) {
-                continue;
-            }
-            Table table = db.getTable(mvId.getId());
-            if (table == null) {
-                continue;
-            }
-            newMvs.add(table);
-            mvs.add((MaterializedView) table);
-        }
-        getRelatedMvs(maxLevel, currentLevel + 1, newMvs, mvs);
-    }
-
-    // get all ref tables within and below root
-    public static List<Table> getAllTables(OptExpression root) {
-        List<Table> tables = Lists.newArrayList();
-        getAllTables(root, tables);
-        return tables;
-    }
-
-    private static void getAllTables(OptExpression root, List<Table> tables) {
-        if (root.getOp() instanceof LogicalScanOperator) {
-            LogicalScanOperator scanOperator = (LogicalScanOperator) root.getOp();
-            tables.add(scanOperator.getTable());
-        } else {
-            for (OptExpression child : root.getInputs()) {
-                getAllTables(child, tables);
-            }
-        }
-    }
-
-    public static boolean isValidMVPlan(OptExpression root) {
-        if (root == null) {
-            return false;
-        }
-        if (isLogicalSPJ(root)) {
-            return true;
-        }
-        if (isLogicalSPJG(root)) {
-            LogicalAggregationOperator agg = (LogicalAggregationOperator) root.getOp();
-            // having is not supported now
-            return agg.getPredicate() == null;
-        }
-        return false;
-    }
-
-    public static boolean isLogicalSPJG(OptExpression root) {
-        if (root == null) {
-            return false;
-        }
-        Operator operator = root.getOp();
-        if (!(operator instanceof LogicalAggregationOperator)) {
-            return false;
-        }
-        LogicalAggregationOperator agg = (LogicalAggregationOperator) operator;
-        if (agg.getType() != AggType.GLOBAL) {
-            return false;
-        }
-
-        OptExpression child = root.inputAt(0);
-        return isLogicalSPJ(child);
-    }
-
-    public static boolean isLogicalSPJ(OptExpression root) {
-        if (root == null) {
-            return false;
-        }
-        Operator operator = root.getOp();
-        if (!(operator instanceof LogicalOperator)) {
-            return false;
-        }
-        if (!(operator instanceof LogicalScanOperator)
-                && !(operator instanceof LogicalProjectOperator)
-                && !(operator instanceof LogicalFilterOperator)
-                && !(operator instanceof LogicalJoinOperator)) {
-            return false;
-        }
-        for (OptExpression child : root.getInputs()) {
-            if (!isLogicalSPJ(child)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public static Pair<OptExpression, LogicalPlan> getRuleOptimizedLogicalPlan(String sql,
-                                                                               ColumnRefFactory columnRefFactory,
-                                                                               ConnectContext connectContext) {
-        StatementBase mvStmt;
-        try {
-            List<StatementBase> statementBases =
-                    com.starrocks.sql.parser.SqlParser.parse(sql, connectContext.getSessionVariable());
-            Preconditions.checkState(statementBases.size() == 1);
-            mvStmt = statementBases.get(0);
-        } catch (ParsingException parsingException) {
-            LOG.warn("parse sql:{} failed", sql, parsingException);
-            return null;
-        }
-        Preconditions.checkState(mvStmt instanceof QueryStatement);
-        Analyzer.analyze(mvStmt, connectContext);
-        QueryRelation query = ((QueryStatement) mvStmt).getQueryRelation();
-        LogicalPlan logicalPlan =
-                new RelationTransformer(columnRefFactory, connectContext).transformWithSelectLimit(query);
-        // optimize the sql by rule and disable rule based materialized view rewrite
-        OptimizerConfig optimizerConfig = new OptimizerConfig(OptimizerConfig.OptimizerAlgorithm.RULE_BASED);
-        optimizerConfig.disableRuleSet(RuleSetType.SINGLE_TABLE_MV_REWRITE);
-        Optimizer optimizer = new Optimizer(optimizerConfig);
-        OptExpression optimizedPlan = optimizer.optimize(
-                connectContext,
-                logicalPlan.getRoot(),
-                new PhysicalPropertySet(),
-                new ColumnRefSet(logicalPlan.getOutputColumn()),
-                columnRefFactory);
-        return Pair.create(optimizedPlan, logicalPlan);
-    }
-
-    public static List<OptExpression> collectScanExprs(OptExpression expression) {
-        List<OptExpression> scanExprs = Lists.newArrayList();
-        OptExpressionVisitor scanCollector = new OptExpressionVisitor<Void, Void>() {
-            @Override
-            public Void visit(OptExpression optExpression, Void context) {
-                for (OptExpression input : optExpression.getInputs()) {
-                    super.visit(input, context);
-                }
-                return null;
-            }
-
-            @Override
-            public Void visitLogicalTableScan(OptExpression optExpression, Void context) {
-                scanExprs.add(optExpression);
-                return null;
-            }
-        };
-        expression.getOp().accept(scanCollector, expression, null);
-        return scanExprs;
-    }
-
-    // get all predicates within and below root
-    public static List<ScalarOperator> getAllPredicates(OptExpression root) {
-        List<ScalarOperator> predicates = Lists.newArrayList();
-        getAllPredicates(root, predicates);
-        return predicates;
-    }
-
-    private static void getAllPredicates(OptExpression root, List<ScalarOperator> predicates) {
-        Operator operator = root.getOp();
-        if (operator.getPredicate() != null) {
-            predicates.add(root.getOp().getPredicate());
-        }
-        if (operator instanceof LogicalJoinOperator) {
-            LogicalJoinOperator joinOperator = (LogicalJoinOperator) operator;
-            if (joinOperator.getOnPredicate() != null) {
-                predicates.add(joinOperator.getOnPredicate());
-            }
-        }
-        for (OptExpression child : root.getInputs()) {
-            getAllPredicates(child, predicates);
-        }
-    }
-
-    public static ScalarOperator canonizePredicate(ScalarOperator predicate) {
-        if (predicate == null) {
-            return null;
-        }
-        ScalarOperatorRewriter rewrite = new ScalarOperatorRewriter();
-        return rewrite.rewrite(predicate, ScalarOperatorRewriter.DEFAULT_REWRITE_SCAN_PREDICATE_RULES);
-    }
-
-    public static ScalarOperator canonizePredicateForRewrite(ScalarOperator predicate) {
-        if (predicate == null) {
-            return null;
-        }
-        ScalarOperatorRewriter rewrite = new ScalarOperatorRewriter();
-        return rewrite.rewrite(predicate, ScalarOperatorRewriter.MV_SCALAR_REWRITE_RULES);
-    }
-
-    public static ScalarOperator getCompensationPredicateForDisjunctive(ScalarOperator src, ScalarOperator target) {
-        List<ScalarOperator> srcItems = Utils.extractDisjunctive(src);
-        List<ScalarOperator> targetItems = Utils.extractDisjunctive(target);
-        int srcLength = srcItems.size();
-        int targetLength = targetItems.size();
-        if (!targetItems.containsAll(srcItems)) {
-            return null;
-        }
-        targetItems.removeAll(srcItems);
-        if (targetItems.isEmpty() && srcLength == targetLength) {
-            // it is the same, so return true constant
-            return ConstantOperator.createBoolean(true);
-        } else if (!targetItems.isEmpty()) {
-            // the target has more or item, so return src
-            return src;
-        } else {
-            return null;
-        }
-    }
-
-    public static boolean isAllEqualInnerJoin(OptExpression root) {
-        Operator operator = root.getOp();
-        if (!(operator instanceof LogicalOperator)) {
-            return false;
-        }
-        if (operator instanceof LogicalJoinOperator) {
-            LogicalJoinOperator joinOperator = (LogicalJoinOperator) operator;
-            boolean isEqualPredicate = isColumnEqualPredicate(joinOperator.getOnPredicate());
-            if (joinOperator.getJoinType() != JoinOperator.INNER_JOIN || !isEqualPredicate) {
-                return false;
-            }
-        }
-        for (OptExpression child : root.getInputs()) {
-            if (!isAllEqualInnerJoin(child)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public static boolean isColumnEqualPredicate(ScalarOperator predicate) {
-        if (predicate == null) {
-            return false;
-        }
-
-        ScalarOperatorVisitor<Boolean, Void> checkVisitor = new ScalarOperatorVisitor<Boolean, Void>() {
-            @Override
-            public Boolean visit(ScalarOperator scalarOperator, Void context) {
-                return false;
-            }
-
-            @Override
-            public Boolean visitCompoundPredicate(CompoundPredicateOperator predicate, Void context) {
-                if (!predicate.isAnd()) {
-                    return false;
-                }
-                for (ScalarOperator child : predicate.getChildren()) {
-                    Boolean ret = child.accept(this, null);
-                    if (!Boolean.TRUE.equals(ret)) {
-                        return false;
-                    }
-                }
-                return true;
-            }
-
-            @Override
-            public Boolean visitBinaryPredicate(BinaryPredicateOperator predicate, Void context) {
-                return predicate.getBinaryType().isEqual()
-                        && predicate.getChild(0).isColumnRef()
-                        && predicate.getChild(1).isColumnRef();
-            }
-        };
-
-        return predicate.accept(checkVisitor, null);
-    }
-
-    public static Map<ColumnRefOperator, ScalarOperator> getColumnRefMap(
-            OptExpression expression, ColumnRefFactory refFactory) {
-        Map<ColumnRefOperator, ScalarOperator> columnRefMap = Maps.newHashMap();
-        if (expression.getOp().getProjection() != null) {
-            columnRefMap.putAll(expression.getOp().getProjection().getColumnRefMap());
-        } else {
-            if (expression.getOp() instanceof LogicalAggregationOperator) {
-                LogicalAggregationOperator agg = (LogicalAggregationOperator) expression.getOp();
-                Map<ColumnRefOperator, ScalarOperator> keyMap = agg.getGroupingKeys().stream().collect(Collectors.toMap(
-                        java.util.function.Function.identity(),
-                        java.util.function.Function.identity()));
-                columnRefMap.putAll(keyMap);
-                columnRefMap.putAll(agg.getAggregations());
-            } else {
-                ColumnRefSet refSet = expression.getOutputColumns();
-                for (int columnId : refSet.getColumnIds()) {
-                    ColumnRefOperator columnRef = refFactory.getColumnRef(columnId);
-                    columnRefMap.put(columnRef, columnRef);
-                }
-            }
-        }
-        return columnRefMap;
-    }
-
-    public static List<ColumnRefOperator> collectScanColumn(OptExpression optExpression) {
-        List<ColumnRefOperator> columnRefOperators = Lists.newArrayList();
-        OptExpressionVisitor visitor = new OptExpressionVisitor<Void, Void>() {
-            @Override
-            public Void visit(OptExpression optExpression, Void context) {
-                for (OptExpression input : optExpression.getInputs()) {
-                    input.getOp().accept(this, input, null);
-                }
-                return null;
-            }
-
-            @Override
-            public Void visitLogicalTableScan(OptExpression optExpression, Void context) {
-                LogicalScanOperator scan = (LogicalScanOperator) optExpression.getOp();
-                columnRefOperators.addAll(scan.getColRefToColumnMetaMap().keySet());
-                return null;
-            }
-        };
-        optExpression.getOp().accept(visitor, optExpression, null);
-        return columnRefOperators;
     }
 }
