+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/rule/AggregateRewriteBaseRule.java
@@ -5,11 +5,11 @@ package com.starrocks.sql.optimizer.rule.transformation.materialization.rule;
 import com.starrocks.sql.optimizer.MaterializationContext;
 import com.starrocks.sql.optimizer.OptExpression;
 import com.starrocks.sql.optimizer.OptimizerContext;
-import com.starrocks.sql.optimizer.Utils;
 import com.starrocks.sql.optimizer.operator.pattern.Pattern;
 import com.starrocks.sql.optimizer.rule.RuleType;
 import com.starrocks.sql.optimizer.rule.transformation.materialization.AggregatedMaterializedViewRewriter;
 import com.starrocks.sql.optimizer.rule.transformation.materialization.MaterializedViewRewriter;
+import com.starrocks.sql.optimizer.rule.transformation.materialization.MvUtils;
 
 public abstract class AggregateRewriteBaseRule extends BaseMaterializedViewRewriteRule {
     protected AggregateRewriteBaseRule(RuleType type, Pattern pattern) {
@@ -18,7 +18,7 @@ public abstract class AggregateRewriteBaseRule extends BaseMaterializedViewRewri
 
     @Override
     public boolean check(OptExpression input, OptimizerContext context) {
-        boolean ret = Utils.isLogicalSPJG(input);
+        boolean ret = MvUtils.isLogicalSPJG(input);
         if (!ret) {
             return false;
         }
